Create controllers for practitioners route.
